Add AudioSpectrum script for music synchronization
This commit introduces a new script, AudioSpectrum.cs. The primary function of this script is to sync in-game activities or visuals with music.
It achieves this by calculating and recalculating audio levels, peak levels, and mean levels at every frame update.
These levels are computed from the audio frequency spectrum data obtained from the AudioListener.
The script also has functionality to save the spectrum data to a file in order to reference it later for debugging or music visualization design improvements.