Add AudioSyncLine script for music visualization
This commit introduces the AudioSyncLine script under Rimaethon/_Scripts/MusicSync. This class extends the functionality of the AudioSyncer class and is dedicated for visualizing music using lines. Lines characteristics (start point, end point, width) and its material can be defined in this class. It also controls how the line reacts to the frequency spectrum of the music by updating line's position according to frequency levels. This will provide a synchronized visual effect to the music played.