Add LineGenerator script for music synchronization
Created a new script called LineGenerator.cs for MusicSync feature. This script will be responsible for creating a line of points between a defined start and end point. This line will be used in musical synchronization, allowing a visual representation of different music tracks. Additionally, an array of points was generated which can be used later for other components in the music sync feature.